fix: stop silently narrowing pow/powi exponents
Integer Number::pow and saturating_pow used exp as u32, silently truncating exponents that do not fit u32 (e.g. 2u64.pow(1<<32) returned 1); they now panic via try_into().expect(). Float::powi and AtomicFloat::fetch_powi took an isize exponent that was narrowed to i32 with as _; they now take i32 directly, matching the standard library.

+++ src/atomic_float.rs
@@ -360,7 +360,7 @@ where
     /// sequence of rounding operations, so the results are not guaranteed
     /// to agree.
     #[cfg(feature = "std")]
-    fn fetch_powi(&self, n: isize, order: Ordering);
+    fn fetch_powi(&self, n: i32, order: Ordering);
 
     /// Atomically raises `self` to a floating point power.
     #[cfg(feature = "std")]
